backend/requirements.txt
asgiref==3.4.1
boto3==1.20.23
botocore==1.23.24
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-cors-headers==3.10.0
django-import-export==2.6.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
flake8==4.0.1
jmespath==0.10.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0